requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
blis==0.7.7
catalogue==2.0.7; python_version >= '3.6'
certifi==2022.5.18.1; python_version >= '3.6'
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.3; python_version >= '3.7'
cycler==0.11.0; python_version >= '3.6'
cymem==2.0.6
fonttools==4.33.3; python_version >= '3.7'
idna==3.3; python_version >= '3'
jinja2==3.1.2; python_version >= '3.7'
kiwisolver==1.4.2; python_version >= '3.7'
langcodes==3.3.0; python_version >= '3.6'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib==3.5.2
murmurhash==1.0.7
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3; python_version >= '3.6'
pandas==1.4.2
pathy==0.6.1; python_version >= '3.6'
pillow==9.1.1
preshed==3.0.6
pydantic==1.8.2; python_full_version >= '3.6.1'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pysqlite3==0.4.7
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.1
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==62.3.2; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smart-open==5.2.1; python_version >= '3.6' and python_version < '4'
spacy-legacy==3.0.9; python_version >= '3.6'
spacy-loggers==1.0.2; python_version >= '3.6'
spacy==3.3.0
sqlalchemy==1.4.36
srsly==2.4.3; python_version >= '3.6'
thinc==8.0.16; python_version >= '3.6'
tqdm==4.64.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typer==0.4.1; python_version >= '3.6'
typing-extensions==4.2.0; python_version >= '3.7'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wasabi==0.9.1
wheel==0.37.1
